requirements.txt
streamlit
pytesseract
Pillow
PyMuPDF
docx2txt
langchain
openai